# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fake Mart API Documentation",
        "description": "This is the API documentation for my Laravel app.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/staff/create": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Create a new staff",
                "description": "Creates a new staff member with all required personal information.",
                "operationId": "createStaff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "position_id",
                                    "name",
                                    "gender",
                                    "date_of_birth",
                                    "place_of_birth",
                                    "address",
                                    "phone",
                                    "nation_id_card"
                                ],
                                "properties": {
                                    "position_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "Male"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-01-01"
                                    },
                                    "place_of_birth": {
                                        "type": "string",
                                        "example": "Phnom Penh"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street, Phnom Penh"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "012345678"
                                    },
                                    "nation_id_card": {
                                        "type": "string",
                                        "example": "N123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "new_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "position_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "Male"
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "example": "1995-01-01"
                                                },
                                                "place_of_birth": {
                                                    "type": "string",
                                                    "example": "Phnom Penh"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Street, Phnom Penh"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "012345678"
                                                },
                                                "nation_id_card": {
                                                    "type": "string",
                                                    "example": "N123456789"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "The name field is required."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/staff/update": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Update staff information",
                "description": "Updates an existing staff record by ID",
                "operationId": "updateStaff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "position_id",
                                    "name",
                                    "gender",
                                    "date_of_birth",
                                    "place_of_birth",
                                    "address",
                                    "phone",
                                    "nation_id_card"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "position_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "Female"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "place_of_birth": {
                                        "type": "string",
                                        "example": "Siem Reap"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "456 Main St, Siem Reap"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "nation_id_card": {
                                        "type": "string",
                                        "example": "N987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "update_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "position_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jane Doe"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "Female"
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "example": "1990-05-15"
                                                },
                                                "place_of_birth": {
                                                    "type": "string",
                                                    "example": "Siem Reap"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "456 Main St, Siem Reap"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "0987654321"
                                                },
                                                "nation_id_card": {
                                                    "type": "string",
                                                    "example": "N987654321"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-06T12:10:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Staff not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "resource not found"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}